algebraic-repair-calculus/proposals/proposal_00/implementation/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc"
version = "0.1.0"
description = "Algebraic Repair Calculus: Three-Sorted Delta Algebra for Provably Correct Pipeline Repair"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ARC Research Team"}
]

dependencies = [
    "sqlglot>=20.0",
    "duckdb>=0.9.0",
    "networkx>=3.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "attrs>=23.0",
    "cattrs>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.80",
    "mypy>=1.5",
]

[project.scripts]
arc = "arc.cli.main:cli"

[tool.setuptools.packages.find]
include = ["arc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-x -q --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
